reorganize typescript cloud functions into modules
